+++ sdk/cpp/core/src/ydk.h
@@ -1,28 +1,26 @@
-//////////////////////////////////////////////////////////////////
-// @file ydk.h
-//
-// YANG Development Kit
-// Copyright 2017 Cisco Systems. All rights reserved
-//
-////////////////////////////////////////////////////////////////
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-//
-//////////////////////////////////////////////////////////////////
+/*  ----------------------------------------------------------------
+ @file ydk.h
+
+ YDK - YANG Development Kit
+ Copyright 2017-2019 Cisco Systems. All rights reserved.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ -------------------------------------------------------------------
+ This file has been modified by Yan Gorelik, YDK Solutions.
+ All modifications in original under CiscoDevNet domain
+ introduced since October 2019 are copyrighted.
+ All rights reserved under Apache License, Version 2.0.
+ ------------------------------------------------------------------*/
 
 #ifndef _YDK_H_
 #define _YDK_H_
@@ -42,7 +40,6 @@ typedef void* OpenDaylightServiceProvider;
 typedef void* Capability;
 typedef void* Repository;
 typedef void* YDKStatePtr;
-
 typedef void* Session;
 
 typedef struct DataNodeChildren
@@ -120,10 +117,18 @@ RootSchemaNode ServiceProviderGetRootSchema(YDKStatePtr, ServiceProvider);
 RootSchemaWrapper ServiceProviderGetRootSchemaNode(YDKStatePtr, ServiceProvider);
 EncodingFormat ServiceProviderGetEncoding(ServiceProvider);
 Session ServiceProviderGetSession(ServiceProvider);
-RootSchemaNode SessionGetRootSchema(YDKStatePtr, Session);
+RootSchemaWrapper SessionGetRootSchemaNode(YDKStatePtr, Session);
 void NetconfServiceProviderFree(ServiceProvider);
 int NetconfServiceProviderGetNumCapabilities(ServiceProvider);
 const char* NetconfServiceProviderGetCapabilityByIndex(ServiceProvider, int);
+
+Session NetconfSessionInit(YDKStatePtr state, Repository repo,
+                           const char * address, const char * username, const char * password, int port,
+                           const char * protocol, boolean on_demand, boolean common_cache, int timeout,
+                           const char * server_certificate_path, const char * private_key_path);
+void NetconfSessionFree(Session);
+
+DataNode SessionExecuteRpc(YDKStatePtr state, Session session, Rpc rpc);
 
 ServiceProvider RestconfServiceProviderInitWithRepo(YDKStatePtr state, Repository repo, const char * address, const char * username, const char * password, int port, EncodingFormat encoding, const char* config_url_root, const char* state_url_root);
 void RestconfServiceProviderFree(ServiceProvider);