+++ sdk/go/core/ydk/path/path_test.go
@@ -1,11 +1,36 @@
+/*  ----------------------------------------------------------------
+ YDK - YANG Development Kit
+ Copyright 2017-2019 Cisco Systems. All rights reserved.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ -------------------------------------------------------------------
+ This file has been modified by Yan Gorelik, YDK Solutions.
+ All modifications in original under CiscoDevNet domain
+ introduced since October 2019 are copyrighted.
+ All rights reserved under Apache License, Version 2.0.
+ ------------------------------------------------------------------*/
+
 package path
 
 import (
 	"github.com/CiscoDevNet/ydk-go/ydk"
-	oc_bgp "github.com/CiscoDevNet/ydk-go/ydk/models/openconfig/bgp"
-	"github.com/CiscoDevNet/ydk-go/ydk/path"
+	"github.com/CiscoDevNet/ydk-go/ydk/types"
+	encoding "github.com/CiscoDevNet/ydk-go/ydk/types/encoding_format"
+	oc_bgp "github.com/CiscoDevNet/ydk-go/ydk/models/ydktest/openconfig_bgp"
 	"github.com/CiscoDevNet/ydk-go/ydk/providers"
 	"testing"
+	"runtime"
+	"path/filepath"
 )
 
 func TestExecuteRpc(t *testing.T) {
@@ -14,10 +39,43 @@ func TestExecuteRpc(t *testing.T) {
 	provider.Connect()
 
 	bgp := oc_bgp.Bgp{}
-	result := path.ExecuteRpc(&provider, &bgp, "ydk:create", "entity", false)
+	data := make(map[string]string)
+	result := provider.ExecuteRpc("create", &bgp, data)
 	if result.Private == nil {
 		t.Error("Operation failed")
+	} else {
+		provider.ExecuteRpc("delete", &bgp, data)
 	}
 
 	provider.Disconnect()
+}
+
+func TestNetconfSessionConnectNoRepo(t *testing.T) {
+	ydk.EnableLogging(ydk.Info)
+	session := NetconfSession{Address: "127.0.0.1", Username: "admin", Password: "admin", Port: 12022}
+	session.Connect()
+
+	rootSchema := session.GetRootSchemaNode()
+
+	bgp := CreateRootDataNode( rootSchema, "openconfig-bgp:bgp")
+	CreateDataNode( bgp, "global/config/as", 65172)
+	bgpCreatePayload := CodecEncode( bgp, encoding.XML, false)
+	ydk.YLogInfo("RPC Payload:\n" + bgpCreatePayload)
+
+	createRpc := CreateRpc( rootSchema, "ydk:create")
+	CreateDataNode( createRpc.Input, "entity", bgpCreatePayload)
+	session.ExecuteRpc(createRpc);
+
+	session.Disconnect()
+}
+
+func TestNetconfSessionConnectWithRepo(t *testing.T) {
+	ydk.EnableLogging(ydk.Debug)
+	_, callerFile, _, _ := runtime.Caller(0)
+	executablePath := filepath.Dir(callerFile)
+	repopath := executablePath + "/../../../../cpp/core/tests/models"
+	repo := types.Repository{Path: repopath}
+	session := NetconfSession{Repo: repo, Address: "127.0.0.1", Username: "admin", Password: "admin", Port: 12022}
+	session.Connect()
+	session.Disconnect()
 }